refactoring: move client side calorie tracker crud to own file
